# openapi.yaml
openapi: 3.0.3
info:
  title: Face Verification Backbone API
  version: "1.0.0"
  description: >
    Identity verification as a service. Two integration models:
    **managed** (we store encrypted templates per tenant) and **stateless**
    (you keep your own data and pass it in). All endpoints except /v1/health
    require an API key sent as the `X-API-Key` header. Verification/compare
    results are HMAC-signed with your tenant's signing secret so your app can
    trust the allow/deny outcome.


    Keys carry a **role**: `admin` (full control) or `verify` (recognition only —
    cannot enrol, delete, or list). Each tenant is isolated. Calls are metered
    per tenant (see `/v1/usage`) and may be subject to a monthly quota
    (HTTP 429 `quota_exceeded`).
servers:
  - url: https://your-host:5000
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ImageOrEmbedding:
      type: object
      description: Either a base64 image or a precomputed 512-d embedding.
      properties:
        image: { type: string, description: "base64 (or data-URL) JPEG/PNG" }
        embedding: { type: array, items: { type: number } }
    Signature:
      type: object
      properties:
        alg: { type: string, example: HMAC-SHA256 }
        ts: { type: string }
        nonce: { type: string }
        hmac: { type: string }
security:
  - ApiKeyAuth: []
paths:
...
  /v1/embed:
    post:
      summary: "Stateless: extract a 512-d face embedding from an image"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties: { image: { type: string } }
      responses:
        "200":
          description: Embedding
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  embedding: { type: array, items: { type: number } }
                  det_score: { type: number }
                  face_px: { type: integer }
                  dims: { type: integer, example: 512 }
  /v1/compare:
    post:
      summary: "Stateless: compare a probe against reference(s), no storage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [probe, references]
              properties:
                probe: { $ref: "#/components/schemas/ImageOrEmbedding" }
                references:
                  type: array
                  items: { $ref: "#/components/schemas/ImageOrEmbedding" }
                threshold: { type: number, description: "default = service match threshold (0.40)" }
      responses:
        "200":
          description: Match result (signed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  match: { type: boolean }
                  best_index: { type: integer }
                  best_score: { type: number }
                  scores: { type: array, items: { type: number } }
                  signature: { $ref: "#/components/schemas/Signature" }
  /v1/enroll:
    post:
      summary: "Managed: enroll a user from one or more images"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string }
                image: { type: string }
                images: { type: array, items: { type: string } }
      responses: { "200": { description: Enrolment summary } }
  /v1/enroll/bulk:
    post:
      summary: "Managed: enrol many users in one call (role: admin)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [people]
              properties:
                people:
                  type: array
                  items:
                    type: object
                    required: [user_id]
                    properties:
                      user_id: { type: string }
                      images: { type: array, items: { type: string } }
                      embeddings: { type: array, items: { type: array, items: { type: number } } }
      responses: { "200": { description: Per-user enrolment results } }
  /v1/verify:
    post:
      summary: "Managed: verify a claimed user (1:1) or identify (1:N)"
      description: >
        Pass `image` for a single-shot check, or `frames` + `token` (from
        /v1/challenge) for an active-liveness head-turn check. Omit `user_id`
        to identify (1:N).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                image: { type: string }
                frames: { type: array, items: { type: string } }
                token: { type: string }
      responses:
        "200":
          description: Verification result (signed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, description: granted }
                  user_id: { type: string }
                  score: { type: number }
                  signature: { $ref: "#/components/schemas/Signature" }
  /v1/identify:
    post:
      summary: "Managed: 1:N identify (image or frames+token)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image: { type: string }
                frames: { type: array, items: { type: string } }
                token: { type: string }
      responses: { "200": { description: Identify result (signed) } }
...
  /v1/users/delete:
    post:
      summary: "Managed: delete one user or many (role: delete)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                user_ids: { type: array, items: { type: string } }
      responses: { "200": { description: Deletion result } }
  /v1/users/export:
    post:
      summary: "Managed: data-subject access — metadata held for a user (role: manage)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties: { user_id: { type: string } }
      responses:
        "200": { description: "What we hold (counts, dim, recent audit) — not the raw template" }
        "404": { description: No record for that user }
  /v1/users/purge:
    post:
      summary: "Managed: erase ALL users in this tenant (role: delete)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties: { confirm: { type: boolean, enum: [true] } }
      responses: { "200": { description: Purge result } }
...